+++ crates/cast/tests/cli/main.rs
@@ -2527,43 +2527,3 @@ Error: Failed to estimate gas: server returned an error response: error code 3: 
 
 "#]]);
 });
-
-forgetest_async!(warns_if_no_transactions_to_broadcast, |prj, cmd| {
-    foundry_test_utils::util::initialize(prj.root());
-
-    prj.add_script(
-        "NoTxScript",
-        r#"
-import "forge-std/Script.sol";
-
-contract NoTxScript is Script {
-    function run() public {
-        vm.startBroadcast();
-        // No real tx created
-        vm.stopBroadcast();
-    }
-}
-"#,
-    )
-    .unwrap();
-
-    cmd.forge_fuse()
-        .args([
-            "script",
-            "--private-key",
-            "0xac0974bec39a17e36ba4a6b4d238ff944bacb478cbed5efcae784d7bf4f2ff80",
-            "--rpc-url",
-            "http://localhost:8545",
-            "--broadcast",
-            "NoTxScript",
-        ])
-        .assert_success()
-        .stdout_eq(str![
-            r#"
-[COMPILING_FILES] with [SOLC_VERSION]
-[SOLC_VERSION] [ELAPSED]
-No files changed, compilation skipped
-No transactions to broadcast
-"#
-        ]);
-});
